CRAPP-35: List approvers of a change request
This commit introduces a new way for computing approvers based on two
mechanisms:
  - usage of the approval right already released in 0.1
  - usage of a new xobject introduced in this commit.

The idea of this xobject is to be able to declare "explicit" approvers
on a page or on a change request: it's possible to declare either users
or groups. If an xobject exists on a change request, and the lists are
not empty, then the approvers are solely based on those lists, and the
approvers defined by those lists don't need to have the approval right.
Now if the lists are empty, the approvers are only based on the approval
right of the pages modified in the change request.
Note that it's possible to also put this xobject in the pages to define
"default approvers" for those pages: this xobject is read when creating
a change request, or when adding changes in a change request, to fill
the actual xobject of the change request.

  * Introduce a new ApproversClass
  * Introduce a new ApproversManager API whose role is to handle
    approvers either for a page or for a change request
  * Provide default implementations for ApproversManager
  * Use this manager in ChangeRequestManager#canMerge and #canReview
  * Use the manager in CreateChangeRequestHandler and
    AddChangesChangeRequestHandler to be sure to fill the xobject of the
change request
  * Add a new script service API to retrieve the lists of approvers
  * Display that list in the change request sheet