CRAPP-48: Allow to create a discussion from a diff line
Those changes setup the new discussions module in change request which
provides the framework to manipulate discussions. It defines several
kind of references to allow attach discussions to different type of
change request elements, and to retrieve them. Right now it's only used
to create discussion attached to a specific diff line, this should
improved in the future for allowing to create other kind of discussions
in the UI.