CRAPP-36: Synchronize page view rights and change request view rights CRAPP-47: Allow to split a change request CRAPP-45: Allow to check that pages are compatible before adding changes
These changes provide a first global mechanism for handling right
synchronization in Change requests between pages and change request.
The idea is that a dedicated component is responsible to check the
consistency of rights between a change request and a new page to be
added, or between a change request and some new changes in the right of
a page already included in a change request. The same component is also
responsible of copying rights when needed.
Then a new global method has
been added in ChangeRequestStorageManager to allow splitting a change
request: it's not entirely finished yet, but the idea is that a change
request containing changes concerning multiple documents can be splitted
in change requests for each document. Splitting a change request have
several impacts like the need to clone filechanges, reviews, or to move
the existing discussions.
Finally a listener has been added which listens on a RightUpdatedEvent
to trigger a consistency check when a change has been performed on a
page handled by a Change request.