CRAPP-86: Allow to create new page as part of a change request
This provide better support of templates when requesting for documents
creation: we check first if the template providers or the template can
be used to not let users perform a creation request if they can't, and
we now properly read the document information from the template when
saving. This also fixes a bug to redirect to the change request when
creating a document from a template.