CRAPP-83: the change request timeline should be more clear about the added messages CRAPP-125: Provide navigation from activity feed CRAPP-91: Allow to reply to any comments
This commit provides a bunch of improvments related to discussions.
First the discussion events are now sent with the messages data. Then
the macro events has been improved to display a specific message
depending on the type of discussion message, and to provide an URL to
allow to navigate to it. Some utilities have been brought to allow to
automatically emphasize a message when a link to it is clicked, and to
allow also to easily navigate through tabs in such case.
Finally the reply to mechanism has been improved to reuse common code,
and to work with diff messages.