CRAPP-138: Keep the list of approvers synchronized between pages and opened change requests
Introduce a new mechanism to ensure that the list of change request
approvers are automatically updated whenever the approver list of a
document is updated.
This change is needed in order to ensure that modifications bring to
list of approvers will impact already opened change requests.
It's particularly important since the approvers mechanism can be updated
with change requests.