CRAPP-139: Allow to delegate approver right to another user
Provide the whole architecture to use delegate approvers. Basically this
works relies on a new dedicated class for delegate approvers, which
might be automatically filled given some properties found in XWikiUsers
class. New configuration options allow to enable the delegate mechanism,
and to chose which properties to watch for filling the delegate
xobjects.
New listeners have been added to watch on the configuration changes or
on the XWikiUsers changes to recompute the delegate xobjects: a job has
been set in place in order to allow computing a large number of users
asynchronously.
Then the review XObject and API has been improved to allow specify an
originalApprover on top of the review author. The strategies have been
improved in order to take into account this originalApprover.
New APIs have been created to check the capability to review given the
delegate mechanism.
Finally the UI has been improved to allow to select on behalf of which
user to perform the review.

WIP:
  - missing doc
  - missing tests
  - missing handling of notifications