CRAPP-299: Approvers are lost when splitting a CR in specific conditions
Clone the filechange and change its type to creation if it's latest
filechange and the published document isn't present anymore, during the
split operation.

This fix is like performing a refresh for this specific type of change,
during the split. It fixes the approvers problem because in such case
we're getting the approvers from the filechange directly.